paypalAPIを使った決済処理を実装、カートの商品一覧から商品削除の際のvalidationを実装

+++ src/htdocs/cart/cart.php
@@ -11,11 +11,11 @@ $data = array();
 $login_name = isset($_SESSION['login_name']) ? $_SESSION['login_name'] : NULL;
 
 //現在のカートの中の購入商品合計数を取得
-$cart_sum_amount_result = get_cart_sum_amount($pdo,$login_name);
+$cart_sum_amount_result = get_cart_sum_amount($pdo, $login_name);
 //現在のカートの中の購入点数を取得
-$purchase_points = get_cart_record($pdo,$login_name);
+$purchase_points = get_cart_record($pdo, $login_name);
 //現在のカートの中の合計金額を取得
-$cart_total_fee = get_cart_total_fee($pdo,$login_name);
+$cart_total_fee = get_cart_total_fee($pdo, $login_name);
 $data['cart_sum_amount_result'] = $cart_sum_amount_result;
 $data['purchase_points'] = $purchase_points;
 $data['cart_total_fee'] = $cart_total_fee;
@@ -27,13 +27,10 @@ $product_id = get_productId_array_from_stock($pdo, $item_id);
 $product_ids = get_productIds($product_id);
 
 $cart_list_info = array();
-foreach ($product_ids as $product_id){
-    array_push($cart_list_info,get_cart_item_info($pdo, $login_name,$product_id));
+foreach ($product_ids as $product_id) {
+    array_push($cart_list_info, get_cart_item_info($pdo, $login_name, $product_id));
 }
 $data['cart_list_info'] = $cart_list_info;
-
-
-
 
 
 $_POST = escape($_POST);
@@ -52,13 +49,20 @@ if ($amount_change) {
     }
 }
 
-if($submit_delete){
-    $result = delete_cart_item($pdo,$submit_delete);
-    if($result === true){
-        header('location:'.CART_PAGE);
-    }
-}
+if ($submit_delete) {
 
+    $check_delete_value = validate_delete_cart_value($pdo, $login_name, $submit_delete);
+
+    if ($check_delete_value === true) {
+        $result = delete_cart_item($pdo, $submit_delete);
+        if ($result === true) {
+            header('location:' . CART_PAGE);
+        }
+    } else {
+        $data['delete_error'] = true;
+    }
+
+}
 
 
 $view = view('/cart.php', $data);
